register functionality now is in auth-service
