+++ src/errors/user-errors.js
@@ -28,28 +28,14 @@ class UsersTableEmptyError extends Error {
     }
 }
 
-class CredentialsTakenError extends Error {
 
-    constructor ( email = null ) {
-        super()
-        this.name = this.constructor.name
-        this.message = 'Ya existe un usuario con esas credenciales'
-
-        if (email){
-         this.message = `Usuario con email: ${email} ya existe.`
-        }
-   
-        this.code = 400
-    }
-  
-}
 
 
 
 module.exports = {
     UserNotFoundError,
     UsersTableEmptyError,
-    CredentialsTakenError,
+ 
 }
